+++ app/(auth)/RedirectingComponent.tsx
@@ -1,0 +1,14 @@
+import CircularProgress from '@mui/material/CircularProgress';
+
+type Params = {
+  path: string;
+};
+export default function RedirectingComponent({ path }: Params) {
+  return (
+    <div>
+      <p>
+        Redirecting to <strong>{path}</strong> <CircularProgress />
+      </p>
+    </div>
+  );
+}
