+++ app/(auth)/signup/page.tsx
@@ -1,0 +1,5 @@
+const SignupPage = () => {
+  return <h1>Signup</h1>;
+};
+
+export default SignupPage;
