+++ app/admin/layout.tsx
@@ -1,0 +1,5 @@
+import AdminLayoutComponent from './AdminLayoutComponent';
+
+export default function AdminLayout({ children }: { children: React.ReactNode }) {
+  return <AdminLayoutComponent>{children}</AdminLayoutComponent>;
+}
