+++ app/admin/lists/page.tsx
@@ -1,0 +1,5 @@
+import ListsComponent from './ListsComponent';
+
+export default function Lists() {
+  return <ListsComponent />;
+}
