+++ app/admin/sermons/page.tsx
@@ -1,0 +1,7 @@
+import AdminSermonsList from './AdminSermonsList';
+
+const AdminSermons = () => {
+  return <AdminSermonsList collectionPath="sermons" />;
+};
+
+export default AdminSermons;
