+++ app/admin/speakers/page.tsx
@@ -1,0 +1,5 @@
+import SpeakersComponent from './SpeakersComponent';
+
+export default function Speakers() {
+  return <SpeakersComponent />;
+}
