+++ app/admin/topics/page.tsx
@@ -1,0 +1,5 @@
+import TopicsComponent from './TopicsComponent';
+
+export default function Topics() {
+  return <TopicsComponent />;
+}
