+++ app/admin/users/page.tsx
@@ -1,0 +1,5 @@
+import UsersComponent from './UsersComponent';
+
+export default function Users() {
+  return <UsersComponent />;
+}
