+++ components/UploadToSubsplashPopup.tsx
@@ -21,7 +21,7 @@ import PopUp from './PopUp';
 import { useCollectionDataOnce } from 'react-firebase-hooks/firestore';
 import { List, listConverter } from '../types/List';
 import ListSelector from './ListSelector';
-import useAuth from '../context/user/UserContext';
+import { useAuth } from '../auth/hooks';
 
 interface UploadToSubsplashPopupProps {
   sermon: Sermon;
@@ -38,7 +38,7 @@ const UploadToSubsplashPopup: FunctionComponent<UploadToSubsplashPopupProps> = (
   setIsUploadingToSubsplash,
   isUploadingToSubsplash,
 }: UploadToSubsplashPopupProps) => {
-  const { user } = useAuth();
+  const { tenant } = useAuth();
   const [autoPublish, setAutoPublish] = useState<boolean>(false);
   const [listArray, setListArray] = useState<List[]>([]);
   const [listArrayFirestore, loading, error] = useCollectionDataOnce(
@@ -102,7 +102,7 @@ const UploadToSubsplashPopup: FunctionComponent<UploadToSubsplashPopupProps> = (
       });
       await updateDoc(sermonRef, {
         status: { ...sermon.status, subsplash: uploadStatus.UPLOADED },
-        approverId: user?.uid,
+        approverId: tenant?.id,
       });
     } catch (error) {
       alert(error);
