+++ components/UserTable.tsx
@@ -17,7 +17,7 @@ import Select from '@mui/material/Select';
 import MenuItem from '@mui/material/MenuItem';
 import { Order, ROLES } from '../context/types';
 import { User } from '../types/User';
-import useAuth from '../context/user/UserContext';
+import { useAuth } from '../auth/hooks';
 
 const stableSort = (array: User[], order: Order, orderBy: keyof User) => {
   function compareEmail(a: User, b: User) {
@@ -143,7 +143,7 @@ const UserTableToolbar = () => {
 const UserTable = (props: { users: User[]; handleRoleChange: (uid: string, role: string) => void }) => {
   const [order, setOrder] = useState<Order>('asc');
   const [orderBy, setOrderBy] = useState<keyof User>('email');
-  const { user: currentUser } = useAuth();
+  const { tenant } = useAuth();
 
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(5);
@@ -192,7 +192,7 @@ const UserTable = (props: { users: User[]; handleRoleChange: (uid: string, role:
                     <TableRow hover tabIndex={-1} key={user.uid}>
                       <TableCell align="center">{user.email}</TableCell>
                       <TableCell align="center">
-                        {currentUser?.uid === user.uid ? (
+                        {tenant?.id === user.uid ? (
                           <Typography>{user.role}</Typography>
                         ) : (
                           <Select
