--- pages/old_index.tsx
@@ -1,43 +1,0 @@
-// /**
-//  * Website's home page
-//  */
-// import { CircularProgress } from '@mui/material';
-// import Box from '@mui/material/Box';
-// import Typography from '@mui/material/Typography';
-// import Stack from '@mui/system/Stack';
-import type { NextPage } from 'next';
-// import Head from 'next/head';
-
-// // Auth
-// import useAuth from '../context/user/UserContext';
-// // Components
-// import LoginPage from './old_login';
-
-const Home: NextPage = () => {
-  //   const { user, loading } = useAuth();
-
-  return (
-    <>
-      {/* //       <Head>
-//         <title>Upper Room Media</title>
-//         <meta property="og:title" content="Upper Room Media" key="title" />
-//         <meta name="description" content="Generated by create next app" />
-//       </Head>
-//       {loading ? (
-//         <Stack justifyContent="center" alignItems="center" width={1} marginTop={2}>
-//           <CircularProgress size={75} />
-//         </Stack>
-//       ) : user ? (
-//         <Stack justifyContent="center" alignItems="center" padding={2}>
-//           <Typography variant="h3">Welcome to the home page</Typography>
-//         </Stack>
-//       ) : (
-//         <Box display="flex" flexDirection="column" justifyContent="space-around" alignItems="center">
-//           <LoginPage></LoginPage>
-//         </Box>
-//       )} */}
-    </>
-  );
-};
-
-export default Home;
