--- pages/old_logout.tsx
@@ -1,16 +1,0 @@
-// /**
-//  * Page for Logout
-//  */
-import type { NextPage } from 'next';
-// import { useEffect } from 'react';
-// import useAuth from '../context/user/UserContext';
-
-const Logout: NextPage = (_props) => {
-  //   const { logoutUser } = useAuth();
-  //   useEffect(() => {
-  //     logoutUser();
-  //   }, []);
-  return <></>;
-};
-
-export default Logout;
